+++ stories/components/illustration/docs/examples/basic-usage.tsx
@@ -1,0 +1,8 @@
+import React from 'react';
+import MailSent from '@semcore/illustration/MailSent';
+
+const Demo = () => {
+  return <MailSent />;
+};
+
+export default Demo;