+++ tools/babel-preset-ui/.babelrc.js
@@ -14,7 +14,6 @@ module.exports = function (babel, opts = {}) {
       [
         '@babel/plugin-transform-runtime',
         {
-          useESModules: true,
           version: require('@babel/runtime/package.json').version,
         },
       ],