+++ semcore/notice/CHANGELOG.md
@@ -2,6 +2,24 @@
 
 CHANGELOG.md standards are inspired by [keepachangelog.com](https://keepachangelog.com/en/1.0.0/).
 
+## [5.44.2] - 2025-02-05
+
+### Changed
+
+- Version patch update due to children dependencies update (`@semcore/animation` [2.40.0 ~> 2.40.1], `@semcore/utils` [4.45.0 ~> 4.45.1], `@semcore/button` [5.41.1 ~> 5.41.2], `@semcore/flex-box` [5.40.0 ~> 5.40.1], `@semcore/icon` [4.57.0 ~> 4.57.1], `@semcore/core` [2.38.0 ~> 2.38.1]).
+
+## [5.44.1] - 2025-02-04
+
+### Added
+
+- Required dependency `@semcore/typeography`.
+
+## [5.44.0] - 2025-02-03
+
+### Changed
+
+- Version minor update due to children dependencies update (`@semcore/animation` [2.39.1 ~> 2.40.0], `@semcore/utils` [4.44.1 ~> 4.45.0], `@semcore/button` [5.40.3 ~> 5.41.0], `@semcore/flex-box` [5.39.1 ~> 5.40.0], `@semcore/icon` [4.55.0 ~> 4.56.0], `@semcore/core` [2.37.1 ~> 2.38.0]).
+
 ## [5.43.3] - 2024-12-23
 
 ### Added