+++ semcore/time-picker/CHANGELOG.md
@@ -2,6 +2,18 @@
 
 CHANGELOG.md standards are inspired by [keepachangelog.com](https://keepachangelog.com/en/1.0.0/).
 
+## [4.49.12] - 2025-02-05
+
+### Changed
+
+- Version patch update due to children dependencies update (`@semcore/button` [5.41.1 ~> 5.41.2], `@semcore/flex-box` [5.40.0 ~> 5.40.1], `@semcore/input` [4.41.0 ~> 4.41.1], `@semcore/select` [4.49.4 ~> 4.49.5], `@semcore/utils` [4.45.0 ~> 4.45.1], `@semcore/core` [2.38.0 ~> 2.38.1]).
+
+## [4.49.11] - 2025-02-05
+
+### Changed
+
+- Version patch update due to children dependencies update (`@semcore/button` [5.41.0 ~> 5.41.1], `@semcore/select` [4.49.3 ~> 4.49.4]).
+
 ## [4.49.10] - 2025-01-21
 
 ### Added